+++ packages/solid/src/components/password-input/password-input.stories.tsx
@@ -8,5 +8,6 @@ export default meta
 
 export { Basic } from './examples/basic'
 export { WithField } from './examples/with-field'
+export { StrengthMeter } from './examples/strength-meter'
 export { WithValidation } from './examples/with-validation'
 export { RootProvider } from './examples/root-provider'