+++ packages/svelte/CHANGELOG.md
@@ -11,6 +11,8 @@ description: All notable changes will be documented in this file.
 - **Field**: Fix issue where `bind:value` doesn't work correctly in `Field.Textarea`, `Field.Input`, and `Field.Select`
   components
 
+- **Password Input**: Fix issue where `bind:value` doesn't work correctly in `PasswordInput.Input` component
+
 ## [5.10.1] - 2025-09-27
 
 ### Fixed